STAGE 5
STAGE 5

DESCRIPTION

This task involves using the Silhouette Score to find the best number of clusters (k). This will help you verify the optimal k you previously identified using the Elbow Method.

The Silhouette Score measures how well-defined your clusters are. A higher score indicates that clusters are dense and well-separated. Your goal is to find the value of k that maximizes this score.

IMPLEMENTATION
You will need to use the sklearn.metrics.silhouette_score function. The process is as follows:

Create a loop where k (the number of clusters) ranges from 2 to 10, inclusive.

Inside the loop, for each value of k:

Create an instance of your CustomKMeans class.

Fit the model to your data (X_full).

Get the predicted cluster labels for the data.

Calculate the score using sklearn.metrics.silhouette_score, passing in your data and the predicted labels.

Store the resulting score.

OBJECTIVES
Print the list of calculated silhouette scores for each value of k.

Plot the silhouette score as a function of k. Use this plot to visually determine the k that corresponds to the highest score.

Compare this optimal k with the one you found previously using the Elbow Method to see if they match.

EXAMPLE OUTPUT:
[9, 8, 7, 6, 5, 4, 3, 2, 1]